refactor: move db-related logic to models
