+++ import-service/lib/import-service-stack.importProductsFile.ts
@@ -2,7 +2,7 @@ import { APIGatewayProxyEventV2, APIGatewayProxyResultV2 } from "aws-lambda";
 import { S3Client, PutObjectCommand } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 
-import { buildResponse, errorResponse } from "../../common/lib/utils";
+import { buildResponse, errorResponse } from "../../common/lib/utils/responses";
 
 export { importProductsFile as handler };
