+++ product-service/test/product-service-stack.getProductsList.test.ts
@@ -1,5 +1,5 @@
 import { handler } from "../lib/product-service-stack.getProductsList";
-import { buildResponse } from "../../common/lib/utils";
+import { buildResponse } from "../../common/lib/utils/responses";
 import { products } from "../data/products";
 import { APIGatewayProxyEventV2 } from "aws-lambda";
